requirements.txt
fastapi
uvicorn
pymongo
python-multipart
yt-dlp
ffmpeg-python
openai-whisper
torch
python-dotenv
yagmail
